requirements.txt
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
flake8==5.0.4
flake8-annotations==2.9.1
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.0
pep8-naming==0.13.2
pluggy==1.4.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytest==7.1.3
python-dotenv==1.0.1
sniffio==1.3.1
soupsieve==2.5
tomli==2.0.1
urllib3==2.2.1
